src/google-cloud-functions/publish-message-to-sns/requirements.txt
boto3
awswrangler
functions-framework==3.1.0
google-cloud-storage
google-cloud-secret-manager